requirements.txt
# Core ML
scikit-learn>=1.3
pandas>=2.0
numpy>=1.24
joblib>=1.3
pyyaml>=6.0

# Data validation
scipy>=1.11

# Visualization
matplotlib>=3.7

# Dev / Testing
pytest>=7.0
pytest-cov>=4.0
ruff>=0.1
pre-commit>=3.0

# Runtime
python-dotenv>=1.0
